PR feedback (for discussion): Separate out virtualization logic into another class
