+++ packages/nimble-components/src/table/index.ts
@@ -9,20 +9,12 @@ import {
     getCoreRowModel as tanStackGetCoreRowModel,
     TableOptionsResolved as TanStackTableOptionsResolved
 } from '@tanstack/table-core';
-import {
-    Virtualizer,
-    VirtualizerOptions,
-    elementScroll,
-    observeElementOffset,
-    observeElementRect,
-    VirtualItem
-} from '@tanstack/virtual-core';
 import type { TableColumn } from '../table-column/base';
 import { TableValidator } from './models/table-validator';
-import { controlHeight } from '../theme-provider/design-tokens';
 import { styles } from './styles';
 import { template } from './template';
 import type { TableRecord, TableRowState, TableValidity } from './types';
+import { TableVirtualizationHelper } from './virtualization-helper';
 
 declare global {
     interface HTMLElementTagNameMap {
@@ -63,37 +55,12 @@ export class Table<
     /**
      * @internal
      */
-    public virtualizer?: Virtualizer<HTMLElement, HTMLElement>;
-
-    /**
-     * @internal
-     */
-    @observable
-    public visibleItems: VirtualItem[] = [];
-
-    /**
-     * @internal
-     */
-    @observable
-    public allRowsHeight = 0;
-
-    /**
-     * @internal
-     */
-    @observable
-    public headerContainerMarginRight = 0;
-
-    /**
-     * @internal
-     */
-    @observable
-    public rowContainerYOffset = 0;
+    public readonly virtualizationHelper: TableVirtualizationHelper<TData>;
 
     private readonly table: TanStackTable<TData>;
     private options: TanStackTableOptionsResolved<TData>;
     private readonly tableInitialized: boolean = false;
     private readonly tableValidator = new TableValidator();
-    private readonly viewportResizeObserver: ResizeObserver;
 
     public constructor() {
         super();
@@ -107,16 +74,8 @@ export class Table<
             autoResetAll: false
         };
         this.table = tanStackCreateTable(this.options);
+        this.virtualizationHelper = new TableVirtualizationHelper(this);
         this.tableInitialized = true;
-        this.viewportResizeObserver = new ResizeObserver(entries => {
-            const borderBoxSize = entries[0]?.borderBoxSize[0];
-            if (borderBoxSize) {
-                // If we have enough rows that a vertical scrollbar is shown, we need to offset the header widths
-                // by the same margin so the column headers align with the corresponding rendered cells
-                const viewportBoundingWidth = borderBoxSize.inlineSize;
-                this.headerContainerMarginRight = viewportBoundingWidth - this.viewport.scrollWidth;
-            }
-        });
     }
 
     public idFieldNameChanged(
@@ -130,12 +89,11 @@ export class Table<
 
     public override connectedCallback(): void {
         super.connectedCallback();
-        this.viewportResizeObserver.observe(this.viewport);
-        this.updateVirtualizer();
+        this.virtualizationHelper.handleConnected();
     }
 
     public override disconnectedCallback(): void {
-        this.viewportResizeObserver.disconnect();
+        this.virtualizationHelper.handleDisconnected();
     }
 
     public dataChanged(
@@ -180,9 +138,7 @@ export class Table<
             };
             return rowState;
         });
-        if (this.$fastController.isConnected) {
-            this.updateVirtualizer();
-        }
+        this.virtualizationHelper.handleRowsUpdated();
     }
 
     private updateTableOptions(
@@ -226,56 +182,6 @@ export class Table<
         });
 
         this.updateTableOptions({ columns: generatedColumns });
-    }
-
-    private updateVirtualizer(): void {
-        const options = this.createVirtualizerOptions();
-        if (this.virtualizer) {
-            this.virtualizer.setOptions(options);
-        } else {
-            this.virtualizer = new Virtualizer(options);
-        }
-        this.virtualizer._willUpdate();
-        this.handleVirtualizerChange();
-    }
-
-    private createVirtualizerOptions(): VirtualizerOptions<
-    HTMLElement,
-    HTMLElement
-    > {
-        const rowHeight = parseFloat(controlHeight.getValueFor(this));
-        return {
-            count: this.tableData.length,
-            getScrollElement: () => {
-                return this.viewport;
-            },
-            estimateSize: (_: number) => rowHeight,
-            enableSmoothScroll: true,
-            overscan: 3,
-            scrollToFn: elementScroll,
-            observeElementOffset,
-            observeElementRect,
-            onChange: () => this.handleVirtualizerChange()
-        } as VirtualizerOptions<HTMLElement, HTMLElement>;
-    }
-
-    private handleVirtualizerChange(): void {
-        const virtualizer = this.virtualizer!;
-        this.visibleItems = virtualizer.getVirtualItems();
-        this.allRowsHeight = virtualizer.getTotalSize();
-        // We're using a separate div ('table-scroll') to represent the full height of all rows, and
-        // the row container's height is only big enough to hold the virtualized rows. So we don't
-        // use the TanStackVirtual-provided 'start' offset (which is in terms of the full height)
-        // to translate every individual row, we just translate the row container.
-        let rowContainerYOffset = 0;
-        if (this.visibleItems.length > 0) {
-            const firstItem = this.visibleItems[0]!;
-            const lastItem = this.visibleItems[this.visibleItems.length - 1]!;
-            if (lastItem.end < this.allRowsHeight) {
-                rowContainerYOffset = firstItem.start - virtualizer.scrollOffset;
-            }
-        }
-        this.rowContainerYOffset = rowContainerYOffset;
     }
 }
